refactor: change setting
