+++ src/app/dashboard/settings/general-details/page.tsx
@@ -1,4 +1,4 @@
-import GeneralDetails from "@/modules/dashboard/pages/geneal-details";
+import GeneralDetails from "@/modules/settings/pages/geneal-details";
 
 export default function Page() {
   return <GeneralDetails />;
